Introduce DocumentConverters
DocumentConverters shepherd Documents back and forth to their associated objects.
