+++ src/wagtail_vector_index/index/base.py
@@ -1,53 +1,86 @@
 from collections.abc import Generator, Iterable
 from dataclasses import dataclass
-from typing import Generic
+from typing import ClassVar, List, Protocol
 
 from django.conf import settings
 
 from wagtail_vector_index.ai import get_chat_backend, get_embedding_backend
+from wagtail_vector_index.ai_utils.backends.base import BaseEmbeddingBackend
 from wagtail_vector_index.backends import get_vector_backend
-
-from ..ai_utils.backends.base import BaseChatBackend, BaseEmbeddingBackend
-from ..base import Document, VectorIndexableType
 
 
 @dataclass
-class QueryResponse(Generic[VectorIndexableType]):
+class Document:
+    """Representation of some content that is passed to vector storage backends.
+
+    A document is usually a part of a model, e.g. some content split out from
+    a VectorIndexedMixin model. One model instance may have multiple documents.
+    """
+
+    id: str
+    vector: List[float]
+    metadata: dict
+
+
+class DocumentConverter(Protocol):
+    def to_documents(
+        self, object: object, *, embedding_backend: BaseEmbeddingBackend
+    ) -> Generator[Document, None, None]:
+        ...
+
+    def from_document(self, document: Document) -> object:
+        ...
+
+    def bulk_to_documents(
+        self, objects: Iterable[object], *, embedding_backend: BaseEmbeddingBackend
+    ) -> Generator[Document, None, None]:
+        ...
+
+    def bulk_from_documents(
+        self, documents: Iterable[Document]
+    ) -> Generator[object, None, None]:
+        ...
+
+
+@dataclass
+class QueryResponse:
     """Represents a response to the VectorIndex `query` method,
     including a response string and a list of sources that were used to generate the response
     """
 
     response: str
-    sources: Iterable[VectorIndexableType]
+    sources: Iterable[object]
 
 
-class VectorIndex(Generic[VectorIndexableType]):
+class VectorIndex:
     """Base class for a VectorIndex, representing some set of documents that can be queried"""
 
-    embedding_backend: BaseEmbeddingBackend
-    chat_backend: BaseChatBackend
-    object_type: type[VectorIndexableType]
+    # The alias of the backend to use for generating embeddings when documents are added to this index
+    embedding_backend_alias: ClassVar[str] = "default"
+    # The alias of the backend (vector database) to use for storing and querying the index
+    vector_backend_alias: ClassVar[str] = "default"
+    # The type of object that this index contains - this is used to convert between documents and objects
+    converter_class: ClassVar[type[DocumentConverter]]
 
     def __init__(
         self,
-        *,
-        chat_backend_alias="default",
-        embedding_backend_alias="default",
-        vector_backend_alias="default",
     ):
         super().__init__()
 
-        self.embedding_backend = get_embedding_backend(embedding_backend_alias)
-        self.chat_backend = get_chat_backend(chat_backend_alias)
-        self.vector_backend = get_vector_backend(alias=vector_backend_alias)
+        self.embedding_backend = get_embedding_backend(self.embedding_backend_alias)
+        self.vector_backend = get_vector_backend(self.vector_backend_alias)
+
         self.backend_index = self.vector_backend.get_index(self.__class__.__name__)
 
     def get_documents(self) -> Iterable[Document]:
         raise NotImplementedError
 
+    def get_converter(self) -> DocumentConverter:
+        return self.converter_class()
+
     def query(
-        self, query: str, *, sources_limit: int = 5
-    ) -> QueryResponse[VectorIndexableType]:
+        self, query: str, *, sources_limit: int = 5, chat_backend_alias: str = "default"
+    ) -> QueryResponse:
         """Perform a natural language query against the index, returning a QueryResponse containing the natural language response, and a list of sources"""
         try:
             query_embedding = next(self.embedding_backend.embed([query]))
@@ -57,7 +90,7 @@ class VectorIndex(Generic[VectorIndexableType]):
         similar_documents = self.backend_index.similarity_search(query_embedding)
 
         sources = self._deduplicate_list(
-            self.object_type.bulk_from_documents(similar_documents)
+            self.get_converter().bulk_from_documents(similar_documents)
         )
 
         merged_context = "\n".join(doc.metadata["content"] for doc in similar_documents)
@@ -70,15 +103,15 @@ class VectorIndex(Generic[VectorIndexableType]):
             merged_context,
             query,
         ]
-        response = self.chat_backend.chat(user_messages=user_messages)
+        chat_backend = get_chat_backend(chat_backend_alias)
+        response = chat_backend.chat(user_messages=user_messages)
         return QueryResponse(response=response.text(), sources=sources)
 
-    def similar(
-        self, object: VectorIndexableType, *, include_self: bool = False, limit: int = 5
-    ) -> list[VectorIndexableType]:
+    def similar(self, object, *, include_self: bool = False, limit: int = 5) -> list:
         """Find similar objects to the given object"""
-        object_documents: Generator[Document, None, None] = object.to_documents(
-            embedding_backend=self.embedding_backend
+        converter = self.get_converter()
+        object_documents: Generator[Document, None, None] = converter.to_documents(
+            object, embedding_backend=self.embedding_backend
         )
         similar_documents = []
         for document in object_documents:
@@ -87,11 +120,11 @@ class VectorIndex(Generic[VectorIndexableType]):
             )
 
         return self._deduplicate_list(
-            self.object_type.bulk_from_documents(similar_documents),
+            converter.bulk_from_documents(similar_documents),
             exclusions=None if include_self else [object],
         )
 
-    def search(self, query: str, *, limit: int = 5) -> list[VectorIndexableType]:
+    def search(self, query: str, *, limit: int = 5) -> list:
         """Perform a search against the index, returning only a list of matching sources"""
         try:
             query_embedding = next(self.embedding_backend.embed([query]))
@@ -103,20 +136,20 @@ class VectorIndex(Generic[VectorIndexableType]):
 
         # Eliminate duplicates of the same objects.
         return self._deduplicate_list(
-            self.object_type.bulk_from_documents(similar_documents)
+            self.get_converter().bulk_from_documents(similar_documents)
         )
 
     @staticmethod
     def _deduplicate_list(
-        documents: Iterable[VectorIndexableType],
+        objects: Iterable[object],
         *,
-        exclusions: Iterable[VectorIndexableType] | None = None,
-    ) -> list[VectorIndexableType]:
+        exclusions: Iterable | None = None,
+    ) -> list:
         if exclusions is None:
             exclusions = []
         # This code assumes that dict.fromkeys preserves order which is
         # behavior of the Python language since version 3.7.
-        return list(dict.fromkeys(item for item in documents if item not in exclusions))
+        return list(dict.fromkeys(item for item in objects if item not in exclusions))
 
     def rebuild_index(self) -> None:
         """Build the index from scratch"""
